requirements.txt
# Core dependencies
#apache-airflow>=2.9.2
#pandas>=2.0.0
#google-generativeai>=0.3.0
#requests>=2.31.0

# For data handling
#python-dotenv>=1.0.0

# For type hints
#typing-extensions>=4.5.0

apache-airflow
pandas
Faker
pyarrow
streamlit
google-generativeai
requests